rename lines_of_code_by_author -> lines_of_code_added_by_author
This concerns
- the name of the output text file
- naming of local variables when writing out to text file

As for documentation:
lines_of_code_added_by_author describes the cumulative amount of code added
by author. It does not take into account any deletions of an author.
If a commit only consists of deletions, then this will not change the
value of 'lines_of_code_added_by_author' as no lines have been added to
the code.