make separator between columns dependent on name format of authors
Up to this commit results were written as comma separated values to text
files. If, however, author names are formatted as
'<last name>, <first name>', this will lead to mismatched data when read
as a comma separated list. If names are formatted using commas, ' | '
will be used as a separator during read and write instead.